+++ CHANGELOG.md
@@ -4,6 +4,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.0] - 2019-08-11
+### Added
+- Channel global tests
+
+### Changed
+- Migrate Consumer start, run and stop methods to async
+
+### Fixed
+- Consumer attributes queue and filter_size to public
+- Channel start and stop methods
+- Channels methods Cython compliance
+
 ## [1.0.12] - 2019-08-09
 ### Changed
 - PyDoc fixes